+++ src/module/Hooks.ts
@@ -3,6 +3,7 @@ import { warn, error, debug, i18n } from '../VariantEncumbrance';
 import {
   getGame,
   VARIANT_ENCUMBRANCE_DFREDS_CONVENIENT_EFFECTS_MODULE_NAME,
+  VARIANT_ENCUMBRANCE_DF_QUALITY_OF_LIFE_MODULE_NAME,
   VARIANT_ENCUMBRANCE_FLAG,
   VARIANT_ENCUMBRANCE_INVENTORY_PLUS_MODULE_NAME,
   VARIANT_ENCUMBRANCE_ITEM_COLLECTION_MODULE_NAME,
@@ -23,6 +24,7 @@ export let invPlusActive;
 export let itemContainerActive;
 export let dfredsConvenientEffectsActive;
 export let invMidiQol;
+export let dfQualityLifeActive;
 
 // export const effectInterface = new EffectInterface();
 
@@ -43,13 +45,13 @@ export const readyHooks = async () => {
         const actorEntity = <Actor>getGame().actors?.get(actorObject.actor._id);
 
         let encumbranceData;
-        if (hasProperty(actorObject.data, `flags.${VARIANT_ENCUMBRANCE_FLAG}.${EncumbranceFlags.DATA}`)) {
-          encumbranceData = <EncumbranceData>getProperty(actorObject.data,`flags.${VARIANT_ENCUMBRANCE_FLAG}.${EncumbranceFlags.DATA}`);
-        }
+        // if (hasProperty(actorObject.data, `flags.${VARIANT_ENCUMBRANCE_FLAG}.${EncumbranceFlags.DATA}`)) {
+        //   encumbranceData = <EncumbranceData>getProperty(actorObject.data,`flags.${VARIANT_ENCUMBRANCE_FLAG}.${EncumbranceFlags.DATA}`);
+        // }
         if(!encumbranceData){
           // const itemsCurrent = <Item[]>actorEntity.data.items.contents;//actorObject.items;// STRANGE BUG actorEntity.data.items.contents
           const actorEntityCurrent = <ActorData>actorObject.actor; // STRANGE BUG <Actor>getGame().actors?.get(actorObject.actor._id);
-          mergeObject(actorEntity.data.flags, actorEntityCurrent.flags);
+          mergeObject(<any>actorEntity.data.flags[VARIANT_ENCUMBRANCE_MODULE_NAME], <any>actorEntityCurrent.flags[VARIANT_ENCUMBRANCE_MODULE_NAME]);
           // mergeObject(actorEntity.data.items, actorObject.items);
           encumbranceData = VariantEncumbranceImpl.calculateEncumbrance(
             actorEntity,
@@ -64,9 +66,25 @@ export const readyHooks = async () => {
           encumbranceElements = htmlElement.find('.encumbrance')[0]?.children;
         }
 
-        const displayedUnits = getGame().settings.get('dnd5e', 'metricWeightUnits')
-          ? <number>getGame().settings.get(VARIANT_ENCUMBRANCE_MODULE_NAME, 'unitsMetric')
-          : <number>getGame().settings.get(VARIANT_ENCUMBRANCE_MODULE_NAME, 'units');
+        let displayedUnits = getGame().settings.get('dnd5e', 'metricWeightUnits')
+          ? <string>getGame().settings.get(VARIANT_ENCUMBRANCE_MODULE_NAME, 'unitsMetric')
+          : <string>getGame().settings.get(VARIANT_ENCUMBRANCE_MODULE_NAME, 'units');
+
+        // Integration with DragonFlagon Quality of Life, Vehicle Cargo Capacity Unit Feature
+        if (dfQualityLifeActive && actorObject.isVehicle && hasProperty(actorObject.data, `flags.${VARIANT_ENCUMBRANCE_DF_QUALITY_OF_LIFE_MODULE_NAME}.unit`)) {
+          const dfVehicleUnit = <number>getProperty(actorObject.data,`flags.${VARIANT_ENCUMBRANCE_DF_QUALITY_OF_LIFE_MODULE_NAME}.unit`);
+          if(dfVehicleUnit){
+            let oldLabel = '';
+            switch (dfVehicleUnit) {
+              case 2240: oldLabel = 'L.Ton'; break;
+              case 2000: oldLabel = 'S.Ton'; break;
+              case 1: oldLabel = 'lbs'; break;
+            }
+            if(oldLabel){
+              displayedUnits = oldLabel;
+            }
+          }
+        }
 
         if (
           !encumbranceElements &&
@@ -205,6 +223,10 @@ export const readyHooks = async () => {
         data?.flags &&
         hasProperty(data, `flags.${VARIANT_ENCUMBRANCE_INVENTORY_PLUS_MODULE_NAME}`)
       ) {
+        doTheUpdate = true;
+      }
+      // Check change on the cargo property of vehicle
+      if (data?.data?.attributes.capacity?.cargo){
         doTheUpdate = true;
       }
 
@@ -581,6 +603,7 @@ export const initHooks = () => {
   dfredsConvenientEffectsActive = <boolean>(
     getGame().modules.get(VARIANT_ENCUMBRANCE_DFREDS_CONVENIENT_EFFECTS_MODULE_NAME)?.active
   );
+  dfQualityLifeActive = <boolean>getGame().modules.get(VARIANT_ENCUMBRANCE_DF_QUALITY_OF_LIFE_MODULE_NAME)?.active; 
 
   // effectInterface.initialize();
 };